+++ scripts/deployAll.ts
@@ -1,8 +1,8 @@
 require('custom-env').env(); // eslint-disable-line
 
-import {deployAllContracts} from './deployers/contracts';
-import {deployContractRegistry} from './deployers/registry';
-import {isLocalNetwork} from './utils/helpers';
+import { deployAllContracts } from './deployers/contracts';
+import { deployContractRegistry } from './deployers/registry';
+import { isLocalNetwork } from './utils/helpers';
 
 async function main() {
   const registry = await deployContractRegistry();
@@ -12,9 +12,11 @@ async function main() {
     await deployAllContracts(registry.address, true);
     console.log('...done - local network ready.');
   } else {
-    console.log('\n⚠️⚠️⚠️⚠️⚠️\n',
+    console.log(
+      '\n⚠️⚠️⚠️⚠️⚠️\n',
       `setup registry address in config file and run\n\n> npm run deploy:contracts:${process.env.NODE_ENV}`,
-      '\n\nto finish deployment\n⚠️⚠️⚠️⚠️⚠️\n');
+      '\n\nto finish deployment\n⚠️⚠️⚠️⚠️⚠️\n'
+    );
   }
 
   console.log('Registry:', registry.address);
@@ -22,7 +24,7 @@ async function main() {
 
 main()
   .then(() => process.exit(0))
-  .catch(error => {
+  .catch((error) => {
     console.error(error);
     process.exit(1);
   });