+++ scripts/deployContracts.ts
@@ -1,10 +1,10 @@
 require('custom-env').env(); // eslint-disable-line
 
-import {deployAllContracts} from './deployers/contracts';
+import { deployAllContracts } from './deployers/contracts';
 
 deployAllContracts('', true)
   .then(() => process.exit(0))
-  .catch(error => {
+  .catch((error) => {
     console.error(error);
     process.exit(1);
   });