+++ scripts/deployRegistry.ts
@@ -1,10 +1,10 @@
 require('custom-env').env(); // eslint-disable-line
 
-import {deployContractRegistry} from './deployers/registry';
+import { deployContractRegistry } from './deployers/registry';
 
 deployContractRegistry()
   .then(() => process.exit(0))
-  .catch(error => {
+  .catch((error) => {
     console.error(error);
     process.exit(1);
   });