+++ test/unit/ChainTest.ts
@@ -1,23 +1,22 @@
-import hre, {ethers} from 'hardhat';
-import {expect, use} from 'chai';
-import {BigNumber, Contract, ContractFactory, Signer} from 'ethers';
-import {waffleChai} from '@ethereum-waffle/chai';
-import {deployMockContract} from '@ethereum-waffle/mock-contract';
+import hre, { ethers } from 'hardhat';
+import { expect, use } from 'chai';
+import { BigNumber, Contract, ContractFactory, Signer } from 'ethers';
+import { waffleChai } from '@ethereum-waffle/chai';
+import { deployMockContract } from '@ethereum-waffle/mock-contract';
 
-import {LeafKeyCoder, LeafType, LeafValueCoder} from '@umb-network/toolbox';
+import { LeafKeyCoder, LeafType, LeafValueCoder } from '@umb-network/toolbox';
 
 import SortedMerkleTree from '../../lib/SortedMerkleTree';
 
 import Registry from '../../artifacts/contracts/Registry.sol/Registry.json';
 import Chain from '../../artifacts/contracts/Chain.sol/Chain.json';
-import ValidatorRegistry
-  from '../../artifacts/contracts/ValidatorRegistry.sol/ValidatorRegistry.json';
+import ValidatorRegistry from '../../artifacts/contracts/ValidatorRegistry.sol/ValidatorRegistry.json';
 import StakingBank from '../../artifacts/contracts/StakingBank.sol/StakingBank.json';
 import Token from '../../artifacts/contracts/Token.sol/Token.json';
-import {toBytes32} from '../../scripts/utils/helpers';
-import {mintBlocks} from '../utils';
+import { toBytes32 } from '../../scripts/utils/helpers';
+import { blockTimestamp, mintBlocks } from '../utils';
 
-const {toWei} = hre.web3.utils;
+const { toWei } = hre.web3.utils;
 
 use(waffleChai);
 
@@ -32,6 +31,7 @@ const setup = async () => {
   const contractFactory = new ContractFactory(Chain.abi, Chain.bytecode, owner);
 
   await contractRegistry.mock.getAddress.withArgs(toBytes32('Chain')).returns(ethers.constants.AddressZero);
+
   const contract = await contractFactory.deploy(contractRegistry.address, blockPadding);
 
   return {
@@ -43,15 +43,23 @@ const setup = async () => {
     validatorRegistry,
     stakingBank,
     contract,
-    contractFactory
+    contractFactory,
   };
 };
 
 const inputs: Record<string, Buffer> = {};
 
 const keys = [
-  'eth-eur', 'btc-eur', 'war-eur', 'ltc-eur', 'uni-eur',
-  'eth-usd', 'btc-usd', 'war-usd', 'ltc-usd', 'uni-usd',
+  'eth-eur',
+  'btc-eur',
+  'war-eur',
+  'ltc-eur',
+  'uni-eur',
+  'eth-usd',
+  'btc-usd',
+  'war-usd',
+  'ltc-usd',
+  'uni-usd',
 ];
 
 keys.forEach((k, i) => {
@@ -64,11 +72,15 @@ const root = tree.getHexRoot();
 const prepareData = async (
   signer: Signer,
   blockHeight: number,
+  dataTimestamp: number,
   root: string | null,
   fcdKeys: string[] = [],
   fcdValues: number[] = []
 ) => {
-  let testimony = ethers.utils.defaultAbiCoder.encode(['uint256', 'bytes32'], [blockHeight, root]);
+  let testimony = ethers.utils.defaultAbiCoder.encode(
+    ['uint256', 'uint256', 'bytes32'],
+    [blockHeight, dataTimestamp, root]
+  );
 
   for (let i = 0; i < fcdKeys.length; i++) {
     testimony += ethers.utils.defaultAbiCoder.encode(['bytes32', 'uint256'], [fcdKeys[i], fcdValues[i]]).slice(2);
@@ -78,36 +90,35 @@ const prepareData = async (
   const affidavit = ethers.utils.arrayify(hashForSolidity);
 
   const sig = await signer.signMessage(affidavit);
-  const {r, s, v} = ethers.utils.splitSignature(sig);
+  const { r, s, v } = ethers.utils.splitSignature(sig);
 
-  return {testimony, affidavit, sig, r, s, v, hashForSolidity};
+  return { testimony, affidavit, sig, r, s, v, hashForSolidity, dataTimestamp };
 };
 
 describe('Chain', () => {
-  let owner: Signer, validator: Signer, validatorAddress: string,
-    contractRegistry: Contract, validatorRegistry: Contract, stakingBank: Contract,
-    contract: Contract, contractFactory: ContractFactory;
+  let owner: Signer,
+    validator: Signer,
+    validatorAddress: string,
+    contractRegistry: Contract,
+    validatorRegistry: Contract,
+    stakingBank: Contract,
+    contract: Contract,
+    contractFactory: ContractFactory;
 
-  const mockSubmit = async (
-    leader = validator,
-    numberOfValidators = 1,
-    totalSupply = 1000,
-    balance = 1000
-  ) => {
+  const mockSubmit = async (leader = validator, totalSupply = 1000, balance = 1000) => {
     await contractRegistry.mock.requireAndGetAddress
-      .withArgs(toBytes32('ValidatorRegistry')).returns(validatorRegistry.address);
+      .withArgs(toBytes32('ValidatorRegistry'))
+      .returns(validatorRegistry.address);
 
-    await validatorRegistry.mock.getNumberOfValidators.returns(numberOfValidators);
-    await validatorRegistry.mock.addresses.returns(await leader.getAddress());
     await contractRegistry.mock.requireAndGetAddress.withArgs(toBytes32('StakingBank')).returns(stakingBank.address);
     await stakingBank.mock.totalSupply.returns(totalSupply);
     await stakingBank.mock.balanceOf.withArgs(await leader.getAddress()).returns(balance);
   };
 
-  const executeSubmit = async (blockHeight: number) => {
+  const executeSubmit = async (blockHeight: number, dataTimestamp: number) => {
     await mockSubmit();
-    const {r, s, v} = await prepareData(validator, blockHeight, root);
-    await contract.connect(validator).submit(root, [], [], [v], [r], [s]);
+    const { r, s, v } = await prepareData(validator, blockHeight, dataTimestamp, root);
+    await contract.connect(validator).submit(dataTimestamp, root, [], [], [v], [r], [s]);
   };
 
   beforeEach(async () => {
@@ -119,7 +130,7 @@ describe('Chain', () => {
       validatorRegistry,
       stakingBank,
       contract,
-      contractFactory
+      contractFactory,
     } = await setup());
   });
 
@@ -135,25 +146,39 @@ describe('Chain', () => {
     it('expect to blockHeight to be 0', async () => {
       expect(await contract.getBlockHeight()).to.eq(0);
     });
+
+    it('not throw on status', async () => {
+      await contractRegistry.mock.requireAndGetAddress
+        .withArgs(toBytes32('ValidatorRegistry'))
+        .returns(validatorRegistry.address);
+
+      await validatorRegistry.mock.getNumberOfValidators.returns(0);
+      await contractRegistry.mock.requireAndGetAddress.withArgs(toBytes32('StakingBank')).returns(stakingBank.address);
+      await stakingBank.mock.totalSupply.returns(0);
+
+      await expect(contract.getStatus()).not.to.be.reverted;
+    });
   });
 
   describe('setBlockPadding()', () => {
     it('expect to set block padding', async () => {
       await expect(contract.setBlockPadding(9))
-        .to.emit(contract, 'LogBlockPadding').withArgs(await owner.getAddress(), 9);
+        .to.emit(contract, 'LogBlockPadding')
+        .withArgs(await owner.getAddress(), 9);
 
       expect(await contract.blockPadding()).to.eq(9);
     });
 
     it('expect to throw when call from NOT an owner', async () => {
-      await expect(contract.connect(validator).setBlockPadding(9))
-        .to.revertedWith('revert Ownable: caller is not the owner');
+      await expect(contract.connect(validator).setBlockPadding(9)).to.revertedWith(
+        'revert Ownable: caller is not the owner'
+      );
     });
   });
 
   describe('recoverSigner()', () => {
     it('expect to return signer', async () => {
-      const {sig, affidavit, r, s, v, hashForSolidity} = await prepareData(validator, 0, root);
+      const { sig, affidavit, r, s, v, hashForSolidity } = await prepareData(validator, 0, 1, root);
 
       const signer = await contract.recoverSigner(hashForSolidity, v, r, s);
 
@@ -164,25 +189,29 @@ describe('Chain', () => {
 
   describe('bytesToBytes32Array()', () => {
     it('expect to convert bytes to array of bytes32', async () => {
-      const arr = ['1', '2', '3'].map(str => `0x${str.padStart(64, '0')}`);
-      const bytes = '0x' + arr.map(item => item.slice(2)).join('');
+      const arr = ['1', '2', '3'].map((str) => `0x${str.padStart(64, '0')}`);
+      const bytes = '0x' + arr.map((item) => item.slice(2)).join('');
       expect(await contract.bytesToBytes32Array(bytes, 0, arr.length)).to.eql(arr);
     });
 
     it('expect to slice bytes by offset', async () => {
-      const arr = ['1', '2', '3'].map(str => `0x${str.padStart(64, '0')}`);
-      const bytes = '0x' + arr.map(item => item.slice(2)).join('');
+      const arr = ['1', '2', '3'].map((str) => `0x${str.padStart(64, '0')}`);
+      const bytes = '0x' + arr.map((item) => item.slice(2)).join('');
       expect(await contract.bytesToBytes32Array(bytes, 2, 1)).to.eql([arr[2]]);
     });
   });
 
   describe('.decodeLeafToNumber()', () => {
     [
-      '0', '1', '2', '999', Number.MAX_SAFE_INTEGER.toString(10),
+      '0',
+      '1',
+      '2',
+      '999',
+      Number.MAX_SAFE_INTEGER.toString(10),
       // @todo not supported yet max uint256
       // '115792089237316195423570985008687907853269984665640564039457584007913129639935',
       // '0xFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFFF'
-    ].forEach(data => {
+    ].forEach((data) => {
       it(`expect ${data} to be decoded properly to number`, async () => {
         const bytes = LeafValueCoder.encode(data, LeafType.TYPE_INTEGER);
         const result = await contract.decodeLeafToNumber(bytes);
@@ -192,20 +221,23 @@ describe('Chain', () => {
 
     describe('throws when', () => {
       it('bytes are empty', async () => {
-        await expect(contract.decodeLeafToNumber('0x00'))
-          .to.revertedWith('revert invalid leaf bytes - missing type metadata');
+        await expect(contract.decodeLeafToNumber('0x00')).to.revertedWith(
+          'revert invalid leaf bytes - missing type metadata'
+        );
       });
 
       it('missing metadata type marker', async () => {
-        await expect(contract.decodeLeafToNumber('0x1111111111'))
-          .to.revertedWith('revert invalid leaf - missing type marker');
+        await expect(contract.decodeLeafToNumber('0x1111111111')).to.revertedWith(
+          'revert invalid leaf - missing type marker'
+        );
       });
 
       it('invalid type', async () => {
         const bytes = LeafValueCoder.encode('1', LeafType.TYPE_FLOAT);
 
-        await expect(contract.decodeLeafToNumber(bytes))
-          .to.revertedWith('revert invalid leaf - invalid type - expect 02:int');
+        await expect(contract.decodeLeafToNumber(bytes)).to.revertedWith(
+          'revert invalid leaf - invalid type - expect 02:int'
+        );
       });
     });
   });
@@ -218,8 +250,8 @@ describe('Chain', () => {
       '0.123456789': `123456789${'0'.repeat(9)}`,
       '0.000000000000001': `1${'0'.repeat(3)}`,
       '999': toWei('999', 'ether'),
-      '0.900719925474099': '900719925474099000'
-    }).forEach(data => {
+      '0.900719925474099': '900719925474099000',
+    }).forEach((data) => {
       it(`expect ${data[0]} to be decoded correctly to float: ${data[1]}`, async () => {
         const bytes = LeafValueCoder.encode(data[0], LeafType.TYPE_FLOAT);
         const result = await contract.decodeLeafToFloat(bytes);
@@ -229,65 +261,86 @@ describe('Chain', () => {
 
     describe('throws when', () => {
       it('bytes are empty', async () => {
-        await expect(contract.decodeLeafToFloat('0x00'))
-          .to.revertedWith('revert invalid leaf bytes - missing type metadata');
+        await expect(contract.decodeLeafToFloat('0x00')).to.revertedWith(
+          'revert invalid leaf bytes - missing type metadata'
+        );
       });
 
       it('missing metadata type marker', async () => {
-        await expect(contract.decodeLeafToFloat('0x1111111111'))
-          .to.revertedWith('revert invalid leaf - missing type marker');
+        await expect(contract.decodeLeafToFloat('0x1111111111')).to.revertedWith(
+          'revert invalid leaf - missing type marker'
+        );
       });
 
       it('invalid type', async () => {
         const bytes = LeafValueCoder.encode('1', LeafType.TYPE_INTEGER);
 
-        await expect(contract.decodeLeafToFloat(bytes))
-          .to.revertedWith('revert invalid leaf - invalid type - expect 03:float');
+        await expect(contract.decodeLeafToFloat(bytes)).to.revertedWith(
+          'revert invalid leaf - invalid type - expect 03:float'
+        );
       });
     });
   });
 
   describe('.getLeaderIndex()', () => {
-    [1, 2, 3, 4].forEach(numberOfValidators => {
+    [1, 2, 3, 4].forEach((numberOfValidators) => {
       it(`expect to return valid index for ${numberOfValidators}`, async () => {
-        const id = (await contract.getLeaderIndex(numberOfValidators,
-          await ethers.provider.getBlockNumber())).toNumber();
+        const id = (
+          await contract.getLeaderIndex(numberOfValidators, await ethers.provider.getBlockNumber())
+        ).toNumber();
 
-        expect(await contract.getLeaderIndex(numberOfValidators,
-          await ethers.provider.getBlockNumber())).to.eq(id, 'round #1');
+        expect(await contract.getLeaderIndex(numberOfValidators, await ethers.provider.getBlockNumber())).to.eq(
+          id,
+          'round #1'
+        );
         await mintBlocks(blockPadding);
-        expect(await contract.getLeaderIndex(numberOfValidators,
-          await ethers.provider.getBlockNumber())).to.eq((id + 1) % numberOfValidators, 'round #2');
+        expect(await contract.getLeaderIndex(numberOfValidators, await ethers.provider.getBlockNumber())).to.eq(
+          (id + 1) % numberOfValidators,
+          'round #2'
+        );
         await mintBlocks(blockPadding);
-        expect(await contract.getLeaderIndex(numberOfValidators,
-          await ethers.provider.getBlockNumber())).to.eq((id + 2) % numberOfValidators, 'round #3');
+        expect(await contract.getLeaderIndex(numberOfValidators, await ethers.provider.getBlockNumber())).to.eq(
+          (id + 2) % numberOfValidators,
+          'round #3'
+        );
         await mintBlocks(blockPadding);
-        expect(await contract.getLeaderIndex(numberOfValidators,
-          await ethers.provider.getBlockNumber())).to.eq((id + 3) % numberOfValidators, 'round #4');
+        expect(await contract.getLeaderIndex(numberOfValidators, await ethers.provider.getBlockNumber())).to.eq(
+          (id + 3) % numberOfValidators,
+          'round #4'
+        );
       });
     });
 
     describe('when block was minted', () => {
       beforeEach(async () => {
-        await executeSubmit(0);
+        await executeSubmit(0, await blockTimestamp());
       });
 
-      [1, 2, 3, 4].forEach(numberOfValidators => {
+      [1, 2, 3, 4].forEach((numberOfValidators) => {
         it(`expect to return valid index for ${numberOfValidators}`, async () => {
-          const id = (await contract.getLeaderIndex(numberOfValidators,
-            await ethers.provider.getBlockNumber())).toNumber();
+          const id = (
+            await contract.getLeaderIndex(numberOfValidators, await ethers.provider.getBlockNumber())
+          ).toNumber();
 
-          expect(await contract.getLeaderIndex(numberOfValidators,
-            await ethers.provider.getBlockNumber())).to.eq(id, 'round #1');
+          expect(await contract.getLeaderIndex(numberOfValidators, await ethers.provider.getBlockNumber())).to.eq(
+            id,
+            'round #1'
+          );
           await mintBlocks(blockPadding);
-          expect(await contract.getLeaderIndex(numberOfValidators,
-            await ethers.provider.getBlockNumber())).to.eq((id + 1) % numberOfValidators, 'round #2');
+          expect(await contract.getLeaderIndex(numberOfValidators, await ethers.provider.getBlockNumber())).to.eq(
+            (id + 1) % numberOfValidators,
+            'round #2'
+          );
           await mintBlocks(blockPadding);
-          expect(await contract.getLeaderIndex(numberOfValidators,
-            await ethers.provider.getBlockNumber())).to.eq((id + 2) % numberOfValidators, 'round #3');
+          expect(await contract.getLeaderIndex(numberOfValidators, await ethers.provider.getBlockNumber())).to.eq(
+            (id + 2) % numberOfValidators,
+            'round #3'
+          );
           await mintBlocks(blockPadding);
-          expect(await contract.getLeaderIndex(numberOfValidators,
-            await ethers.provider.getBlockNumber())).to.eq((id + 3) % numberOfValidators, 'round #4');
+          expect(await contract.getLeaderIndex(numberOfValidators, await ethers.provider.getBlockNumber())).to.eq(
+            (id + 3) % numberOfValidators,
+            'round #4'
+          );
         });
       });
     });
@@ -297,25 +350,40 @@ describe('Chain', () => {
     describe('without FCD', () => {
       it('expect to mint a block', async () => {
         await mockSubmit();
-        const {r, s, v} = await prepareData(validator, 0, root);
+        const { r, s, v, dataTimestamp } = await prepareData(validator, 0, await blockTimestamp(), root);
 
-        await expect(contract.connect(validator).submit(root, [], [], [v], [r], [s])).not.to.be.reverted;
+        await expect(contract.connect(validator).submit(dataTimestamp, root, [], [], [v], [r], [s])).not.to.be.reverted;
         console.log(await contract.blocks(0));
       });
 
       it('fail when signature do not match', async () => {
         await mockSubmit();
-        const {r, s, v} = await prepareData(validator, 0, root);
+        const { r, s, v, dataTimestamp } = await prepareData(validator, 0, await blockTimestamp(), root);
 
-        const differentRoot = '0x00000000000000000000000000000000000000000000000000000000000001';
-        await expect(contract.connect(validator).submit(differentRoot, [], [], [v], [r], [s])).to.be.reverted;
+        const differentRoot = '0x0000000000000000000000000000000000000000000000000000000000000001';
+        await expect(contract.connect(validator).submit(dataTimestamp, differentRoot, [], [], [v], [r], [s]))
+          .to.be // recover sig will return some other address than validator, so mock function will fail
+          .revertedWith('Mock on the method is not initialized');
+      });
+
+      it('fail when data older than 25 minutes', async () => {
+        await mockSubmit();
+        const t = await blockTimestamp();
+        const { r, s, v, dataTimestamp } = await prepareData(validator, 0, t - 25 * 60 - 1, root);
+
+        await expect(contract.connect(validator).submit(dataTimestamp, root, [], [], [v], [r], [s])).to.be.revertedWith(
+          'data are older than 25 minutes'
+        );
       });
 
       describe('when block submitted', () => {
+        let previousDataTimestamp: number;
+
         beforeEach(async () => {
           await mockSubmit();
-          const {r, s, v} = await prepareData(validator, 0, root);
-          await contract.connect(validator).submit(root, [], [], [v], [r], [s]);
+          const { r, s, v, dataTimestamp } = await prepareData(validator, 0, await blockTimestamp(), root);
+          previousDataTimestamp = dataTimestamp;
+          await contract.connect(validator).submit(dataTimestamp, root, [], [], [v], [r], [s]);
         });
 
         it('expect to blockHeight NOT change when minimal padding not reached', async () => {
@@ -352,6 +420,16 @@ describe('Chain', () => {
           expect(fcds.timestamp).to.gt(0);
         });
 
+        it('fail when data NOT newer than previous block', async () => {
+          await mintBlocks(blockPadding);
+          await mockSubmit();
+          const { r, s, v, dataTimestamp } = await prepareData(validator, 1, previousDataTimestamp, root);
+
+          await expect(
+            contract.connect(validator).submit(dataTimestamp, root, [], [], [v], [r], [s])
+          ).to.be.revertedWith('can NOT submit older data');
+        });
+
         describe('verify Proof', () => {
           describe('.verifyProofForBlock()', () => {
             it('expect to validate proof for selected key-value pair', async () => {
@@ -381,8 +459,8 @@ describe('Chain', () => {
           it('expect to validate multiple proofs as once', async () => {
             const keys = Object.keys(inputs).slice(-3);
             const blockHeights = new Array(keys.length).fill(0);
-            const {proofs, proofItemsCounter} = tree.getFlatProofsForKeys(keys);
-            const leaves = keys.map(k => tree.getLeafForKey(k));
+            const { proofs, proofItemsCounter } = tree.getFlatProofsForKeys(keys);
+            const leaves = keys.map((k) => tree.getLeafForKey(k));
             const result = new Array(keys.length).fill(true);
 
             expect(await contract.verifyProofs(blockHeights, proofs, proofItemsCounter, leaves)).to.eql(result);
@@ -392,9 +470,10 @@ describe('Chain', () => {
         describe('when still on the same block height', () => {
           it('expect submit to be reverted', async () => {
             await mockSubmit();
-            const {r, s, v} = await prepareData(validator, 0, root);
-            await expect(contract.connect(validator).submit(root, [], [], [v], [r], [s]))
-              .to.revertedWith('revert block already mined for current blockHeight');
+            const { r, s, v, dataTimestamp } = await prepareData(validator, 0, await blockTimestamp(), root);
+            await expect(
+              contract.connect(validator).submit(dataTimestamp, root, [], [], [v], [r], [s])
+            ).to.revertedWith('revert block already mined for current blockHeight');
           });
 
           describe('when minimal padding reached', () => {
@@ -409,15 +488,16 @@ describe('Chain', () => {
             describe('when block mined for new block height', () => {
               beforeEach(async () => {
                 await contract.setBlockPadding(1);
-                await executeSubmit(1);
+                await executeSubmit(1, await blockTimestamp());
                 await contract.setBlockPadding(100);
               });
 
               it('expect to revert when submit again for same block', async () => {
                 await mockSubmit();
-                const {r, s, v} = await prepareData(validator, 1, root);
-                await expect(contract.connect(validator).submit(root, [], [], [v], [r], [s]))
-                  .to.revertedWith('revert block already mined for current blockHeight');
+                const { r, s, v, dataTimestamp } = await prepareData(validator, 1, await blockTimestamp(), root);
+                await expect(
+                  contract.connect(validator).submit(dataTimestamp, root, [], [], [v], [r], [s])
+                ).to.revertedWith('revert block already mined for current blockHeight');
               });
             });
           });
@@ -432,10 +512,18 @@ describe('Chain', () => {
       describe('when block submitted', () => {
         beforeEach(async () => {
           await mockSubmit();
-          const {r, s, v} = await prepareData(validator, 0, root, fcdKeys, fcdValues);
+          const { r, s, v, dataTimestamp } = await prepareData(
+            validator,
+            0,
+            await blockTimestamp(),
+            root,
+            fcdKeys,
+            fcdValues
+          );
 
-          await expect(contract.connect(validator).submit(root, fcdKeys, fcdValues, [v], [r], [s]))
-            .to.emit(contract, 'LogMint');
+          await expect(
+            contract.connect(validator).submit(dataTimestamp, root, fcdKeys, fcdValues, [v], [r], [s])
+          ).to.emit(contract, 'LogMint');
         });
 
         it('expect to get First Class Data', async () => {
@@ -456,19 +544,52 @@ describe('Chain', () => {
     });
   });
 
+  it('expect to getStatus()', async () => {
+    await mintBlocks(blockPadding);
+    await mockSubmit();
+    let { r, s, v, dataTimestamp } = await prepareData(validator, 0, await blockTimestamp(), root);
+    await contract.connect(validator).submit(dataTimestamp, root, [], [], [v], [r], [s]);
+
+    await mintBlocks(blockPadding);
+    await mockSubmit();
+    ({ r, s, v, dataTimestamp } = await prepareData(validator, 1, dataTimestamp + 1, root));
+    await contract.connect(validator).submit(dataTimestamp, root, [], [], [v], [r], [s]);
+
+    await contractRegistry.mock.getAddress.withArgs(toBytes32('ValidatorRegistry')).returns(validatorRegistry.address);
+    await validatorRegistry.mock.getNumberOfValidators.returns(1);
+    await validatorRegistry.mock.addresses.withArgs(0).returns(validatorAddress);
+    await validatorRegistry.mock.validators.withArgs(validatorAddress).returns(validatorAddress, 'abc');
+    await contractRegistry.mock.getAddress.withArgs(toBytes32('StakingBank')).returns(stakingBank.address);
+    await stakingBank.mock.totalSupply.returns(123);
+    await stakingBank.mock.balanceOf.withArgs(validatorAddress).returns(321);
+
+    await mintBlocks(blockPadding);
+
+    const status = await contract.getStatus();
+
+    expect(status.lastDataTimestamp).to.eq(dataTimestamp, 'invalid lastDataTimestamp');
+    expect(status.lastBlockHeight).to.eq(1, 'invalid block height');
+    expect(status.nextBlockHeight).to.eq(2, 'invalid block height');
+    expect(status.nextLeader).to.eq(validatorAddress, 'invalid validator');
+    expect(status.validators).to.eql([validatorAddress], 'invalid validators list');
+    expect(status.powers).to.eql([BigNumber.from(321)], 'invalid powers');
+    expect(status.locations).to.eql(['abc'], 'invalid locations');
+    expect(status.staked).to.eq(123, 'invalid staked');
+  });
+
   describe('update/replace contract', () => {
     let newChain: Contract;
 
     beforeEach(async () => {
       await mockSubmit();
-      let {r, s, v} = await prepareData(validator, 0, root);
-      await contract.connect(validator).submit(root, [], [], [v], [r], [s]);
+      let { r, s, v, dataTimestamp } = await prepareData(validator, 0, await blockTimestamp(), root);
+      await contract.connect(validator).submit(dataTimestamp, root, [], [], [v], [r], [s]);
 
       await mintBlocks(blockPadding);
 
       await mockSubmit();
-      ({r, s, v} = await prepareData(validator, 1, root));
-      await contract.connect(validator).submit(root, [], [], [v], [r], [s]);
+      ({ r, s, v, dataTimestamp } = await prepareData(validator, 1, dataTimestamp + 1, root));
+      await contract.connect(validator).submit(dataTimestamp, root, [], [], [v], [r], [s]);
 
       await contractRegistry.mock.getAddress.withArgs(toBytes32('Chain')).returns(contract.address);
       newChain = await contractFactory.deploy(contractRegistry.address, blockPadding);